Remove loadbalancer metrics from neutron
